edison: add support for oneshot boot targets
Read the boot target from the osnib for oneshoot reboots. Should
actually read the oshob address from the scu and then examine the
structure of the oshob to find the osnib pointer. Also a different
field of the osnib is used than the conventional boot_target,
which seems to not take updates from linux presently.

Signed-off-by: Scott D Phillips <scott.d.phillips@intel.com>
Change-Id: I842c2087c5d928948ee55afe22cc85d97e321783
Reviewed-on: https://android.intel.com/420447