fb_mmc: Change gpt flashing to take gpt fragments instead of full gpt
The gpt flashing target now takes a serialized data structure that
specifies the gpt entries instead of a full mbr and gpt table. The
chief benefit of this approach is that the actual lba offsets of
the partitions are determined on the device and can adapt to the
exact size of the flash part.

Signed-off-by: Scott D Phillips <scott.d.phillips@intel.com>
Change-Id: I2d5546ff64d8f68de18867f605d42e1f18b5196a
Reviewed-on: https://android.intel.com/423418