boot_brillo: Move metadata location and implement recovery try counts
Previously the boot control metadata was located at a fixed offset
of the misc partition beyond the bootloader_message structure.
Here it is relocated to within the 'slot_suffix' field of the
bootloader_message structure.

Also logic to boot into a recovery partition is added with
appropriate try count logic as specified in the latest Brillo Boot
Loader Requirements.

Signed-off-by: Scott D Phillips <scott.d.phillips@intel.com>
Change-Id: I3e88bbfa8c0a93b1afed1bc91830a22a38df0129
Reviewed-on: https://android.intel.com/424073