HACK: mmc: only do 8KiB aligned multiblock writes
For edison we are seeing multiblock writes that silently only
partially complete. It's not clear exactly what the alignment
requirement is and I haven't seen any code in the linux kernel for
tangier that applies logic like this, so I'm not sure why this
issue isn't seen there.

Here we add a hack where multiblock writes are only done when the
start address and the length are 8KiB aligned. Other writes are
handled as a series of single block writes.

Change-Id: I03f1e1f3b34a2d511d2c87b309121c285079bc5e
Signed-off-by: Scott D Phillips <scott.d.phillips@intel.com>