fastboot: Add flashing lock/unlock command
This command changes the device state flag from the security partition.
After this change the flash command will work only if the device is in
unlocked state.
The switch between lock <-> unlock device state requires erasing userdata.
User must have the possibility of deciding whether they want to continue
with the flashing command or not.

Change-Id: Idc4397f7a73a398f21a583c90b9d4b8af71d4623
Tracked-On: https://jira01.devtools.intel.com/browse/BP-316
Signed-off-by: Alexandru Timohi <alexandru.timohi@intel.com>
Signed-off-by: Irina Patru <irina.patru@intel.com>
Signed-off-by: Mihai Serban <mihai.serban@intel.com>
Reviewed-on: https://android.intel.com/476197