Add bootloader_message_ab struct
Move A/B-specific fields, effectively slot_suffix as in this CL, out of
bootloader_message struct. Such A/B-specific fields are expected to be
managed by the A/B-bootloader or boot control HAL, which shouldn't be
cleared by recovery or uncrypt.

Round up the bootloader_message struct to 2-KiB. Round up the
bootloader_message_ab struct to 4-KiB.

Based on: https://android-review.googlesource.com/#/c/237711
Change-Id: Ibcead3c353ee65bc68e00f528ed753ece806b215
Signed-off-by: Sanrio Alvares <sanrio.alvares@intel.com>
Reviewed-on: https://android.intel.com/514218
Reviewed-by: Stanacar, Stefan <stefan.stanacar@intel.com>
Tested-by: Stanacar, Stefan <stefan.stanacar@intel.com>